Fixes crash in deleteTebDetours()
When a detour is deleted, the iterator may now be passed the last element of
the container, yet the code proceded to dereference it to test whether teb
planning had failed. Since the iterator is no longer valid in this case, it
could SIGSEGV. The right thing to do is continue the loop at this point
so that the while() condition can be retested.