Badger Specific: Disable deletion of detour Teb's
When the Badger BAR robot faces a narrow gap, or an "penensula", the best path
may require turning around, but this detour deletion can cause planning to
fail or for us to oscillate with a different homotopy class. So, we'll disable
it for now. The only cost is a small performance hit in some edge cases.