Visualize planning failures
Publishes a second set of visualization topics under the "failure"
namespace within Teb's namespace. When planning fails, publish the
failed path and teb state, etc here for display in Rviz.

Fixed for compatibility with upstream change to HCP which removes the
"best" teb if the feasibility check fails by keeping a reference to it
as infeasible_teb_.