Fix SIGSEGV if main thread is interrupted by boost
The problem was that boost::thread_group::join_all() is a boost interruption
point, so will exit early with an exception if someone requests that this
thread be interrupted. But if that happens, then the thread(s) we were
waiting on continues running. On next planner iteration, we may create a new
thread to operate on the same TEB object, leading to data corruption.

The solution is simple: Disable thread interruption during optimizeAllTEBs().