WIP-Experiments with improved motion control
We have been experiencing underdamped control as a result of inadequate
translation of the planned path into twist messages.

This quick hack improves it substantially, especially in the linear component
by using a more up-to-date current position of the robot in the twist
calculation. It had been using the position as it was before teb
optimization, which can be a number of milliseconds old.

This change doesn't entirely fix the problem, especially in the angular
dimension. Further improvement could take the robot's velocity into account
(forward-projecting its position a little bit to account for system
latencies). Or we could run a control loop to calculate new twists at a
much higher frequency than we run the full teb planning loop. Finally, we
could even add an edge to the optimization graph to give the first few poses
of the path some affinity for their existing positions, once the optimizer has
stabilized.

Currently, this hack contains layering violations and extraneous debug code.