HCP: Allow a tunable number of TEBs in same class as "best"
Prior to this change, excess loops could develop in the plan. Sometimes
the initial teb would turn 270-degrees to the left rather than 90
degrees to the right, especially when the goal is close to an obstacle.
In other cases, if a dynamic obstacle moved through the scene, it could
latch a 360-degree loop in the plan. Because such looped plans are a
local minimum, the optimizer is unable to escape from them.

A simple way to demonstrate this problem is to use a teleop joystick to
cause the robot to turn in place while it is navigating. Each
turn-in-place adds a loop to the plan so that the robot would "unwind"
all of these loops as it proceeds to the goal!

This change introduces a new parameter
max_number_plans_in_current_class. When set to 2 or more, it will allow
the HomotopyClassPlanner to accept multiple candidate plans that are in
the same class as the current "best" plan. This gives a chance for
exploration to add a new plan in the same homotopy class as the current
one, so that it can discover a more optimal plan without the loops.

NOTE: The actual number of homotopy classes being considered is reduced
by (max_number_plans_in_current_class - 1), so one should generally
increase max_number_classes whenever this parameter is increased. It
makes no sense for this parameter to be greater than max_number_classes.