HCP: Randomly drop non-best candidate TEBs
Corrects a problem with the Homotopy Class Planner in which it can
become "fixated" on reevaluating the same few alternate TEBs each
planning cycle because the list of tebs is full to the configured
max_number_classes.

For example, if max_number_classes is 5, all 5 slots in the tebs_ array
will quickly be filled with the current plan + 4 alternatives. Then,
each planning cycle will not try any new classes via exploration since
the tebs_ list is already full. It remains in this state Until the robot
moves far enough or the scene changes enough to cause one or more of the
4 alternates to become invalid.  This can prevent the the HCP from
discovering a better path in a timely manner.

This commit introduces a simple fix in which each of the non-best plans
in the list will be randomly dropped with a probability controlled by
the selection_dropping_probability parameter, allowing exploration to
add new candidate(s) to the list.