Make all trajectory parameters dynamic reconfigurable
The code handles on-the-fly changes to these parameters just like all
the others, so go ahead and make them dynamic:

* min_samples
* max_samples
* global_plan_prune_distance
* min_resolution_collision_check_angular
* control_look_ahead_poses

This is in preparation for adding a new parameter related to the last
one in this list. It would be awkward for it to by dynamic while this is
not.