Add inflation_cost_exponent param for nonlinear cost
As with the main obstacle cost (based on min_obstacle_dist), a large
inflation distance leads to instability in path optimization due to the
sum of the straight-line inflation costs of obstacles on opposite sides of the
trajectory being a horizontal line (with zero derivative). Again, the solution is to make this cost function non-linear.

Add new parameter inflation_cost_exponent. When set to any positive
number, this will make the inflation cost nonlinear. The entire cost
curve is lower than the linear cost, so a corresponding increase in
inflation_weight may be needed.

Augments: 70382cb8eb ("Add nonlinear part to obstacle cost to improve
                       narrow gap [behavior]")