h_signature: Revert to power function for f0 (numerator)
A previous bugfix for Issue #90 corrected the denominator of Al which
was generally being left equal to (1, 0) before that fix. After that
fix, the denominator generally becomes very large if there are more than
a few obstacles in the scene (e.g. With ~30 obstacles, the denominator
was around 10^17)

In what was likely an earlier attempt to fix that caused by the
incorrect denominator calculation, the original power function used in
the numerator (f0) had been replaced by a simple multiplication.
However, now that we are calculating the denominator correctly while
still using only a small numerator, all the H-signatures have become
extremely small (on the order of 10^-16 with around 30 obstacles in my
tests).  Thus, ::isEqual() method always returns true so we never
explore any alternate plans, rendering the HCP ineffective.

This change restores the original calculation of f0 thus making it large
enough that the huge denominator will not dengenerate the H-signature to
near zero. In my tests, this makes the HCP function correctly again.

Fixes: 682ad2403fa4 ("bugfix in calculateHSignature. Fixes #90")