h_signature: Use cplx typedef consistently
Some parts of the code use the typedef, and some declare the equivalent
type directly or that of its value_type, making it impractical to
ever change the typedef.

Move the typedef to the abstract base class so that it is visible
everywhere that it is needed. Replace all direct declarations of it or
its value_type with references to the typedef.

Motivation: Performance. Currently, this is log double (80-bit floating
point on amd_64 architecture). The H-signature calculation speeds up by
almost 2x when changing this type to a regular double.

Not actually changing the type here but just cleaning up the code in
preparation. Actually changing the type likely requires improvements to
calculateHSignature() to reduce potential for numerical instability.