h_signature: Undo recent changes for speed
First, this upstream change corrected a bug in the denominator
calculation of Al, but broke the H-signature by making them degenerate
to near-zero values for non-trivial obstacle count:

682ad2403fa4 ("bugfix in calculateHSignature. Fixes #90")

Then, my commit fixes the numerator to be a power function (as it was in
an original commented out line of code), making the signatures
functional again:

9bb9cf5c86714 ("h_signature: Revert to power function for f0 (numerator)")

The result of both of those commits is functional but the code executes
about 5 to 10 times slower than before either of those changes!
Although the math seems wrong, in practice, the code worked sufficiently
correctly before those fixes.

This commit effectively reverts both of them for performance reasons.

TODO:
* Research why this seems to work even though it differs from the
  research paper substantially.
* Research alternative ways to improve performance, such as is
  pre-calculating the denominator of Al and re-using for all obstacles.