Introduce new control law that is much more precise
The current control law (i.e. planner_->getVelocityCommand()) suffers
from serveral drawbacks:

1. If control_look_ahead_poses is 1 (which was the only option until
recently), it is subject to fishtailing about the path and / or taking
curves way too wide as planner keeps moving the path outward as the
robot lags the curve.

2. If control_look_ahead_poses is 3 or more to reduce fishtailing,
then it will cut corners way too close since the control law turns
toward the angle of the carrot pose regardless of what its bearing
is from the robot.

Introducing a new control law based on the course "Control of Mobile
Robots" by Dr. Magnus Egerstedt of the Georgia Institute of Technology,
Section 7.4, "A Clever Trick."

A simple linear equation directly transforms the direction of the chosen
carrot pose into a twist command (vx, omega) which results in a "control
point" a short distance in front of the robot's axis of rotation moving
immediately in the direction of the carrot. The orientation of the
carrot is ignored until we reach the final goal.

The result is very precise path following.

This control law was tested by freezing planning and showing that a
physical differential drive robot would follow the plan with
sub-centimeter deviation, no fishtailing, no overshoot and smooth
accelerations, even at high speeds.  In addition, it recovers quickly
and smoothly from a disturbance caused by teleop'ing off the plan track.

Because this controller does not assume that the robot is always at
"Pose #0" of the plan, it lays some of the ground work to be able to run
the TEB planner at a lower frequency than the controller, which could
greatly reduce CPU usage. However, doing that does not fit very well in
the architecture of move_base, it is deferred.