add optional left-right obstacle inflation edges
Add ability to optionally add left-right inflation edges to one or
more of the pose verteces. The left-right edges are inflation-only
edges where only obstacles to the left or right are taken into
account. The starting and stopping pose vertex are configurable, as
is the pose skip for adding edges. Also, the left and right exponents
and inflation distances are separate so an asymetry can be created
that is useful to cause the base to navigate on one side or other of
a corridor or aisle.

Currently this is only implemented when using Costmap3D.