Preadjust 1/3: Add ability for a VertexPose to have fixed theta
Needed to allow the "preadjust goal if blocked" recovery to work. This
recovery moves the pose but is restrained from changing its orientation.

This is a bit of a hack, needed because g2o does not allow you to fix
only one dimension of a vertex, and the inheritance hierarchy makes it
impractical to accomplish using a subclass or a templatized class with
one fewer dimension.