Preadjust 3/5: Add preadjust_goal_if_blocked recovery
When the goal (either horizon goal or final goal) is in conflict with an
obstacle, pre-adjust it using a small optimzation graph consisting only
of the goal pose, a via point to attract it to the nominal goal, and an
obstacle edge to push it away from obstacles. Its theta is fixed; it can
only move positionally.

Uses caching and hysteresis to keep the solution stable when feasible so
that the robot will not twist & turn excessively near the goal.