add edges for nonlethal obstacles
Costmaps can track obstacles that are inbetween the cost of free
space and lethal space. This can be useful when an object is detected,
such as a cord on the floor or a mat, that can be safely navigated,
but is better to be avoided if possible.

Add options to add edges for non-lethal obstacles. Add edge types
for Costmap3D, where querying the distance to only non-lethal obstacles
is possible.