preadjust: Fix wrong orientation
In case where no free pose is found in existing plan so that we select
the robot's current position as the starting pose for the
pre-adjust-goal-if-blocked, the code neglected to change its orientation
to match that of the nominal goal. In rare cases, this caused us to
finish the nav goal at the wrong orientation. Fixed.

ROS-154