control law: fix slerp_line error around -pi
The slerp_line function calculated an incorrect orientation angle
when one pose is near -pi and the other pose is near +pi.
The deltaTheta variable was not normalized, which caused the
new pose to be interpolated the long way on the arc (around nearly
an entire circle) instead of the short way. Fix this by normalizing
deltaTheta prior to multiplying by the interpolation fraction.