control law: add min angle, general turn-in-place
The previous implementation of the control law special cased a
rotation in place when near the goal, and looked ahead on the path
until the carrot was at least carrot_min_dist away. However, if the
path contains a spiral such that the pose min_carrot_dist away is
significantly different in orientation than the current orientation
the control law poorly follows the plan, sometimes turning the wrong
way.

Deal with this by generalizing the turn-in-place control to activate
any time the carrot pose is too close to the robot pose. Prevent
under-achieving on rotation by projecting forward in time if the
carrot pose is within both a minimum distance and a minimum angle.
Limiting by both distance and angle prevents projecting forward so
far that the angle changes wildly.

When turning in place, if the carrot pose is not at the goal, simply
issue the direct rotational velocity to follow the rotation in the
plan. If the carrot pose is at the goal, use the P controller to
prevent over-shooting the goal.