control law: linearly interpolate angle in slerp_arc
Under certain circumstances, the path gets perturbed such that for a
pose in the path, the pose at the end of the arc from that pose to
the next pose has a significantly different orientation than the next
pose. This can cause the turn-in-place controller to turn the wrong
way, or the main control law to think the path is backwards when it
is forwards. Fix this by always linearly interpolating the angle
between two poses instead of using the tangential pose of the
position along the arc.